refactor : move _getEndOfHeaders and _getEndOfData back to vhdfile
