+++ packages/vhd-lib/src/Vhd/VhdAbstract.js
@@ -1,17 +1,6 @@
-import {
-  BLOCK_UNUSED,
-  FOOTER_SIZE,
-  HEADER_SIZE,
-  PARENT_LOCATOR_ENTRIES,
-  PLATFORM_NONE,
-  SECTOR_SIZE,
-  PLATFORM_W2KU,
-} from '../_constants'
 import { computeBatSize, sectorsRoundUpNoZero, sectorsToBytes } from './_utils'
-import { createLogger } from '@xen-orchestra/log'
+import { PLATFORM_NONE, SECTOR_SIZE, PLATFORM_W2KU } from '../_constants'
 import assert from 'assert'
-
-const { debug } = createLogger('vhd-lib:VhdAbstract')
 
 export class VhdAbstract {
   #header
@@ -123,47 +112,6 @@ export class VhdAbstract {
   // common
   get batSize() {
     return computeBatSize(this.header.maxTableEntries)
-  }
-
-  // Returns the first address after metadata. (In bytes)
-  _getEndOfHeaders() {
-    const { header } = this
-
-    let end = FOOTER_SIZE + HEADER_SIZE
-
-    // Max(end, block allocation table end)
-    end = Math.max(end, header.tableOffset + this.batSize)
-
-    for (let i = 0; i < PARENT_LOCATOR_ENTRIES; i++) {
-      const entry = header.parentLocatorEntry[i]
-
-      if (entry.platformCode !== PLATFORM_NONE) {
-        end = Math.max(end, entry.platformDataOffset + sectorsToBytes(entry.platformDataSpace))
-      }
-    }
-
-    debug(`End of headers: ${end}.`)
-
-    return end
-  }
-
-  // Returns the first sector after data.
-  _getEndOfData() {
-    let end = Math.ceil(this._getEndOfHeaders() / SECTOR_SIZE)
-
-    const sectorsOfFullBlock = this.sectorsOfBitmap + this.sectorsPerBlock
-    const { maxTableEntries } = this.header
-    for (let i = 0; i < maxTableEntries; i++) {
-      const blockAddr = this._getBatEntry(i)
-
-      if (blockAddr !== BLOCK_UNUSED) {
-        end = Math.max(end, blockAddr + sectorsOfFullBlock)
-      }
-    }
-
-    debug(`End of data: ${end}.`)
-
-    return sectorsToBytes(end)
   }
 
   async writeParentLocator({ parentLocatorId, platformCode = PLATFORM_NONE, data = Buffer.alloc(0) }) {
