+++ packages/vhd-lib/src/Vhd/VhdFile.js
@@ -1,4 +1,12 @@
-import { BLOCK_UNUSED, FOOTER_SIZE, HEADER_SIZE, PLATFORM_NONE, PLATFORM_W2KU, SECTOR_SIZE } from '../_constants'
+import {
+  BLOCK_UNUSED,
+  FOOTER_SIZE,
+  HEADER_SIZE,
+  PLATFORM_NONE,
+  PLATFORM_W2KU,
+  SECTOR_SIZE,
+  PARENT_LOCATOR_ENTRIES,
+} from '../_constants'
 import { computeBatSize, sectorsToBytes, buildHeader, buildFooter, BUF_BLOCK_UNUSED } from './_utils'
 import { createLogger } from '@xen-orchestra/log'
 import { fuFooter, fuHeader, checksumStruct } from '../_structs'
@@ -101,12 +109,52 @@ export class VhdFile extends VhdAbstract {
     assert.strictEqual(bytesRead, n)
     return buffer
   }
+  // Returns the first address after metadata. (In bytes)
+  _getEndOfHeaders() {
+    const { header } = this
+
+    let end = FOOTER_SIZE + HEADER_SIZE
+
+    // Max(end, block allocation table end)
+    end = Math.max(end, header.tableOffset + this.batSize)
+
+    for (let i = 0; i < PARENT_LOCATOR_ENTRIES; i++) {
+      const entry = header.parentLocatorEntry[i]
+
+      if (entry.platformCode !== PLATFORM_NONE) {
+        end = Math.max(end, entry.platformDataOffset + sectorsToBytes(entry.platformDataSpace))
+      }
+    }
+
+    debug(`End of headers: ${end}.`)
+
+    return end
+  }
 
   // return the first sector (bitmap) of a block
   _getBatEntry(blockId) {
     const i = blockId * 4
     const blockTable = this.#blockTable
     return i < blockTable.length ? blockTable.readUInt32BE(i) : BLOCK_UNUSED
+  }
+
+  // Returns the first sector after data.
+  _getEndOfData() {
+    let end = Math.ceil(this._getEndOfHeaders() / SECTOR_SIZE)
+
+    const sectorsOfFullBlock = this.sectorsOfBitmap + this.sectorsPerBlock
+    const { maxTableEntries } = this.header
+    for (let i = 0; i < maxTableEntries; i++) {
+      const blockAddr = this._getBatEntry(i)
+
+      if (blockAddr !== BLOCK_UNUSED) {
+        end = Math.max(end, blockAddr + sectorsOfFullBlock)
+      }
+    }
+
+    debug(`End of data: ${end}.`)
+
+    return sectorsToBytes(end)
   }
 
   containsBlock(id) {
